move extension handling tests
